Split transaction tests into separate files
